CNKAF-752: Store and fetch remove broker api state (#1797)
* CNKAF-752: Store and fetch remove broker api state

This change adds the logic to store and fetch databalancer api states
from the api state persistent topic. The change has provision for both
Add broker and remove broker states, but only supports remove broker for
now. The code is based on KafkaBasedLog that is used for other similar
use cases in Kafka (connection configuration, license etc).

Added new unit and integration test to test the persistence store.

+++ ce-sbk/src/main/java/com/linkedin/kafka/cruisecontrol/SbkTopicUtils.java
@@ -156,13 +156,13 @@ public final class SbkTopicUtils {
     public static final class SbkTopicConfig {
 
         public final long minRetentionTimeMs;
-        public final int replicationFactor;
+        public final short replicationFactor;
         public final String topic;
         public final int partitionCount;
         public final String cleanupPolicy;
 
         public SbkTopicConfig(long minRetentionTimeMs,
-                              int replicationFactor,
+                              short replicationFactor,
                               String topic,
                               int partitionCount,
                               String cleanupPolicy) {
@@ -177,7 +177,7 @@ public final class SbkTopicUtils {
     public static final class SbkTopicConfigBuilder {
 
         private long minRetentionTimeMs;
-        private int replicationFactor;
+        private short replicationFactor;
         private String topic;
         private int partitionCount;
         private String cleanupPolicy;
@@ -197,11 +197,11 @@ public final class SbkTopicUtils {
 
         public SbkTopicConfigBuilder setReplicationFactor(Map<String, ?> config,
                                                           String replicationFactorConfig,
-                                                          int defaultReplicationFactor) {
+                                                          short defaultReplicationFactor) {
             String replicationFactor = (String) config.get(replicationFactorConfig);
             this.replicationFactor = replicationFactor == null || replicationFactor.isEmpty()
                     ? defaultReplicationFactor
-                    : Integer.parseInt(replicationFactor);
+                    : Short.parseShort(replicationFactor);
             return this;
         }
